+++ benchmarks/compare/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  env: {
+    commonjs: true,
+    node: true,
+    es2020: true
+  },
+  extends: 'eslint-config-digitalbazaar',
+  root: true
+};